requirements.txt
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
mysqlclient==2.2.0
numpy
oauthlib==3.2.2
pandas
Pillow
pip==23.3.1
platformdirs==3.9.1
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
reportlab
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.8.0
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
twilio==8.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
wheel==0.38.4
wrapt==1.15.0
yarl==1.9.2